rename cluster creation command
